refactor(datasources): hoist root_cause beside sanitize_url, with a hop cap (#431)
The source-chain walker lived module-private in `openqa::client`, where
`describe` uses it to skip a `reqwest::Error`'s Display and report the
rustls/io cause underneath. The send paths of the other datasources need
exactly that for their TLS-detail lines, so it moves to `http.rs` next to
`sanitize_url`, gaining the same self-referential-chain cap
`is_ssl_verification_error` already carries (the openQA copy walked
unbounded; the two caps count slightly differently, which the doc states
rather than papering over).

`ssl_error_detail(&HttpError)` wraps it for the TLS branches. The
signature takes the converted error deliberately: a raw `reqwest::Error`
here would leak through the `to_string` fallback, and `&HttpError` makes
that a compile error instead of a doc plea.

`load_ca_bundle` now converts through `From` instead of naming the
variant, keeping the boundary the single construction path.
